Fix #27 (ReflectionUtil too chatty)
The reflection util was spitting out warning messages even when it
successfully updated a field value. There was no way to get feedback
from the DerivedClassIterator whether it successfully handled a clazz or
not. I've added a feedback mechanism to the DervicedClassIterator,
updated all implementations and made the setFieldValue onlye generate a
warn-level msg when it really failed.